/atlas: add NUTS scripts
Note that this add this gist [1] as a git submodule, so that we can track changes in the output files without adding them directly to this repo, which would make it slow to clone and manage in general.
We'll need a proper solution going forward as gists are supposed to be around ~10 files so this seems to be a stretch, but it's a start.

Closes #35

[1] https://gist.github.com/mindrones/b9538f1b7308d1a2f2d54c927116e825